processing improved

+++ main.py
@@ -12,8 +12,16 @@ from psycopg2 import Error, connect
 
 
 def wlog(vid, code, msg):
-    level = {"i": "INFO", "w": "WARN", "e": "ERROR"}.get(code, "INFO")
-    print("%s %s (%s) %s" % (datetime.now().strftime("%y%m%d:%H%M%S"), vid, level, msg), flush=True)
+    level = {"i": "INFO", "w": "WARN", "e": "ERRO"}.get(code, "INFO")
+    print("%s %s (%s) %s" % (datetime.now().strftime("%y%m%d:%H%M%S"), level, vid, msg), flush=True)
+
+
+def whisper(fn, path=None):
+    m = os.path.join(WHISPER["model"]["path"], fext(fn.rsplit("_", 1)[1], "bin"))
+    cmd = WHISPER["exec"] + [fn.rsplit("_", 2)[1], "-f", fn, "-m", m, "-of", fn[:-4]]
+    if path is not None:
+        cmd[0] = "%s/%s" % (path, cmd[0])
+    return cmd
 
 
 def make_test_env(connector, n):
@@ -31,10 +39,7 @@ def make_test_env(connector, n):
     fn, fn2 = random.sample(fns, 2)
     update_status(connector, get_auid(fn), "processing", datetime.now())
     update_status(connector, get_auid(fn2), "processing", datetime.now())
-
-    cmd = WHISPER["exec"] + [fn.rsplit("_", 2)[1], "-f", fn, "-m", get_model(fn), "-of", fn[:-4]]
-    cmd[0] = "./" + cmd[0]
-    subprocess.Popen(cmd, stdout=open(fext(cmd[10], "log"), "w"), stderr=subprocess.STDOUT)
+    subprocess.Popen(whisper(fn, path="."), stdout=open(fext(fn, "log"), "w"), stderr=subprocess.STDOUT)
 
 
 def fext(fn, ext="wav"):
@@ -43,10 +48,6 @@ def fext(fn, ext="wav"):
 
 def get_auid(fn):
     return fn.split("_", 1)[0][7:]
-
-
-def get_model(fn):
-    return os.path.join(WHISPER["model"]["path"], fext(fn.rsplit("_", 1)[1], "bin"))
 
 
 def get_input(connector):
@@ -58,11 +59,8 @@ def get_input(connector):
                 inp = ["%s/%s_%s_%s.wav" % (it[3].strftime("%y%m%d"), it[0], it[1], it[2]) for it in cur.fetchall()]
     except Error as err:
         wlog("db", "e", str(err).replace("\n", " "))
-        return None
-    n_inp = len(inp)
-    inp = [it for it in inp if os.path.exists(it)]
-    wlog("loaded", "i", "local = %s, total = %s" % (len(inp), n_inp))
-    return inp
+        return list()
+    return [it for it in inp if os.path.exists(it)]
 
 
 def wait_proc(fn, proc, connector):
@@ -80,11 +78,11 @@ def wait_proc(fn, proc, connector):
     return False
 
 
-def run(cmds, connector, it=0, sleep=5.0):
+def run(inp, connector, sleep=5.0):
     ps = list()
-    for cmd in cmds:
-        auid = get_auid(cmd[it])
-        ps.append((cmd[it], subprocess.Popen(cmd, stdout=open(fext(cmd[it], "log"), "w"), stderr=subprocess.STDOUT)))
+    for fn, cmd in inp:
+        auid = get_auid(fn)
+        ps.append((fn, subprocess.Popen(cmd, stdout=open(fext(fn, "log"), "w"), stderr=subprocess.STDOUT)))
         wlog(auid, "i", "processing")
         update_status(connector, auid, "processing", datetime.now())
     while len(ps) > 0:
@@ -92,47 +90,47 @@ def run(cmds, connector, it=0, sleep=5.0):
         ps = [(fn, proc) for (fn, proc) in ps if wait_proc(fn, proc, connector)]
 
 
-def check_proc(connector):
-    ps, fn_pid = os.popen("ps --no-headers --cols 1000 -C whisper -o pid,cmd").readlines(), dict()
-    os.wait()
-    for it in ps:
-        it = it.split()
-        if len(it) >= 16:
-            fn_pid["%s.wav" % it[-1]] = int(it[0])
-
+def check_proc(connector, t_coef=2):
     query = "SELECT auid, lang, model, loaded, processing, attempt FROM " + CONF["table"] + \
             " WHERE processing IS NOT NULL AND failed IS NULL AND success IS NULL"
     try:
         with connector:
             with connector.cursor() as cur:
                 cur.execute(query)
-                ps = cur.fetchall()
+                proc = cur.fetchall()
     except Error as err:
         wlog("db", "e", str(err).replace("\n", " "))
         return 0
+    if not proc:
+        return 0
 
-    o_proc = len(ps) - len(fn_pid)
+    ps, fn_pid = os.popen("ps --no-headers --cols 1000 -C whisper -o pid,cmd").readlines(), dict()
+    os.wait()
     for it in ps:
+        it = it.strip().split()
+        if len(it) >= 16:
+            fn_pid["%s.wav" % it[-1]] = int(it[0])
+
+    for it in proc:
         fn = "%s/%s_%s_%s.wav" % (it[3].strftime("%y%m%d"), it[0], it[1], it[2])
         if os.path.exists(fn):
-            ttl = 2 * (os.path.getsize(fn) * WHISPER["model"][CONF["n_thread"]][it[2]] / 32000.)
+            ttl = t_coef * os.path.getsize(fn) * WHISPER["model"][CONF["n_thread"]][it[2]] / 32000.
             if (datetime.now() - it[4]).total_seconds() > ttl:
                 pid = fn_pid.pop(fn, None)
                 if pid is not None:
                     try:
                         os.kill(pid, 9)
-                        wlog(it[0], "w", "killed: processing too long")
+                        wlog(it[0], "w", "killed")
                         update_status(connector, it[0], "failed", "killed:toolong")
                     except Exception as err:
-                        wlog("processing", "e", str(err).replace("\n", " "))
+                        wlog("pkill", "e", str(err).replace("\n", " "))
                 else:
                     if it[5] == 1:
-                        wlog(it[0], "w", "resumed: attempt started")
+                        wlog(it[0], "w", "attempt resumed")
                         update_status(connector, it[0], "processing", None)
                     else:
-                        wlog(it[0], "w", "killed: repeated attempt failed")
-                        update_status(connector, it[0], "failed", "killed:repeated")
-    wlog("processing", "i", "local = %s of %s, total = %s" % (len(fn_pid), CONF["max_cpu"], o_proc + len(fn_pid)))
+                        wlog(it[0], "w", "attempt failed")
+                        update_status(connector, it[0], "failed", "failed:attempt")
     return CONF["max_cpu"] - len(fn_pid)
 
 
@@ -200,17 +198,15 @@ def update_status(connector, auid, key, val):
 
 if __name__ == "__main__":
     connector = init_connector(CONF)
-    if connector is None:
-        sys.exit(1)
-
-    os.chdir(DATA_DIR)
-    if len(sys.argv) == 2 and sys.argv[1].isdecimal():
-        make_test_env(connector, int(sys.argv[1]))
-    else:
-        n = check_proc(connector)
-        if n < 1:
-            sys.exit(0)
-        run([WHISPER["exec"] + [fn.rsplit("_", 2)[1], "-f", fn, "-m", get_model(fn), "-of", fn[:-4]]
-             for fn in get_input(connector)[:n]], connector, 10)
-
+    if connector is not None:
+        os.chdir(DATA_DIR)
+        if len(sys.argv) == 2 and sys.argv[1].isdecimal():
+            make_test_env(connector, int(sys.argv[1]))
+        else:
+            n = check_proc(connector)
+            if n > 0:
+                inp = get_input(connector)
+                if inp:
+                    wlog("task", "i", "new = %s, queue = %s" % (n, len(inp)))
+                    run([(fn, whisper(fn)) for fn in inp[:n]], connector)
     connector.close()
